edge-security/backend/api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pycryptodome>=3.18.0
python-dotenv>=0.19.0
